\name{nimble2-package}
\alias{nimble2-package}
\alias{nimble2}
\docType{package}
\title{
\packageTitle{nimble2}
}
\description{
\packageDescription{nimble2}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{nimble2}

Maintainer: \packageMaintainer{nimble2}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = nimble2)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{nimble2}}
%\section{Documentation Index}{\packageIndices{nimble2}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
